+++ azure-pipelines.yml
@@ -125,67 +125,3 @@ stages:
       displayName: Publish container logs
       condition: always()
       continueOnError: 'true'
-  - job: Azure
-    pool:
-      vmImage: 'ubuntu-latest'
-    steps:
-    - task: GoTool@0
-      inputs:
-        version: '1.18'
-    - script: |
-        make build
-        cp ./dist/linux_amd64/release/rad ./dist/
-        chmod +x ./dist/rad
-    - script: |
-        az login --service-principal \
-            --username $(INTEGRATION_TEST_SP_APP_ID) \
-            --password $(INTEGRATION_TEST_SP_PASSWORD) \
-            --tenant $(INTEGRATION_TEST_TENANT_ID)
-      env:
-        INTEGRATION_TEST_SP_APP_ID: $(INTEGRATION_TEST_SP_APP_ID)
-        INTEGRATION_TEST_SP_PASSWORD: $(INTEGRATION_TEST_SP_PASSWORD)
-        INTEGRATION_TEST_TENANT_ID: $(INTEGRATION_TEST_TENANT_ID)
-    - bash: |
-        export RANID=$(date +%s%N)
-        arr=("westeurope" "australiaeast")
-        index=$(expr $RANID % 2)
-        export LOCATION=${arr[$index]}
-        echo "##vso[task.setvariable variable=RADENVIRONMENT]rad-$RANID"
-        go run cmd/rad/main.go \
-              env init azure \
-              -e rad-$RANID \
-              --subscription-id ${{ variables.INTEGRATION_TEST_SUBSCRIPTION_ID }} \
-              --resource-group rad-$RANID \
-              --location $LOCATION \
-              --loganalytics-workspace-id "${{ variables.INTEGRATION_TEST_LOGANALYTICS_WORKSPACEID }}" \
-              --deployment-template "${{ variables.RADIUS_DEPLOYMENT_TEMPLATE_LOCATION }}" \
-              --chart ${{ variables.RADIUS_CHART_LOCATION }}  \
-              --image ${{ variables.DOCKER_REGISTRY }}/radius-rp \
-              --tag ${{ variables.REL_VERSION }}
-    - script: |
-        export PATH=$(Build.SourcesDirectory)/dist:$PATH
-        cd $(Build.SourcesDirectory)
-        which rad || { echo "cannot find rad"; exit 1; }
-        rad bicep download
-      displayName: rad bicep download
-    - script: |
-        export PATH=$(Build.SourcesDirectory)/dist:$PATH
-        make test-functional-azure
-      env:
-        INTEGRATION_TEST_SP_APP_ID: $(INTEGRATION_TEST_SP_APP_ID)
-        INTEGRATION_TEST_SP_PASSWORD: $(INTEGRATION_TEST_SP_PASSWORD)
-        INTEGRATION_TEST_TENANT_ID: $(INTEGRATION_TEST_TENANT_ID)
-        TEST_TIMEOUT: ${{ variables.AZURE_FUNCTIONALTEST_TIMEOUT }}
-    - script: |
-        az group delete \
-          --subscription ${{ variables.INTEGRATION_TEST_SUBSCRIPTION_ID }} \
-          --resource-group $(RADENVIRONMENT) \
-          --yes \
-          --verbose \
-          --no-wait
-      displayName: Delete environment resource groups
-      condition: always()
-      continueOnError: 'true'
-        
-        
-
